fix: Convert header and cpp files to hpp files - amongst other changes.
