+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/src/eos_helpers.cpp
@@ -32,8 +32,8 @@
 #include <codecvt>
 #include <eos_types.h>
 
-#include "Config/PlatformConfig.h"
-#include "Config/ProductConfig.h"
+#include "Config/PlatformConfig.hpp"
+#include "Config/ProductConfig.hpp"
 
 /**
   * @brief Retrieves the system cache directory.
